backend/requirements.txt
fastapi
uvicorn
sqlalchemy
rapidfuzz
pydantic
python-multipart
requests
python-dotenv
google-generativeai
